# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChatGPT Account Checker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Get current admin info
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
        "401":
          description: Not authenticated

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
  /admin/keys:
    get:
      operationId: listKeys
      tags: [admin]
      summary: List license keys (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: number }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Keys list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyPage"
    post:
      operationId: createKeys
      tags: [admin]
      summary: Create one or more license keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeysInput"
      responses:
        "201":
          description: Created keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchKeys"
  /admin/keys/{id}:
    get:
      operationId: getKey
      tags: [admin]
      summary: Get a single key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyDetail"
        "404":
          description: Not found
    patch:
      operationId: updateKey
      tags: [admin]
      summary: Revoke / lock / unlock / extend a key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyUpdateInput"
      responses:
        "200":
          description: Updated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyDetail"
    delete:
      operationId: deleteKey
      tags: [admin]
      summary: Revoke a key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"

  # ── Users ─────────────────────────────────────────────────────────────────────
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List Telegram users (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: number }
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
...
  /admin/users/{telegramId}/reset-trial:
    post:
      operationId: resetUserTrial
      tags: [admin]
      summary: Reset a user's free trial count to 0
      parameters:
        - name: telegramId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Trial reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }

  # ── Logs ──────────────────────────────────────────────────────────────────────
  /admin/logs/usage:
    get:
      operationId: listUsageLogs
      tags: [admin]
      summary: List usage logs (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: number }
        - name: telegram_id
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Usage logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageLogPage"
  /admin/logs/audit:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: List audit logs (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: number }
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"
...
  /admin/orders:
    get:
      operationId: listOrders
      tags: [admin]
      summary: List payment orders
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: limit
          schema: { type: number }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersEnvelope"
...
  /admin/inventory/auto-stock:
    post:
      operationId: autoStock
      tags: [admin]
      summary: Auto-create keys to reach the stock target for a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoStockInput"
      responses:
        "200":
          description: Auto-stock result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoStockResult"

  # ── Settings ──────────────────────────────────────────────────────────────────
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      summary: Get system settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsEnvelope"
    put:
      operationId: updateSettings
      tags: [admin]
      summary: Update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
components:
  schemas:
...
    OkResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }

    # ── Checker ───────────────────────────────────────────────────────────────────
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AdminSession:
      type: object
      required: [ok, username]
      properties:
        ok: { type: boolean }
        username: { type: string }
    AdminProfile:
      type: object
      properties:
        admin:
          type: object
          properties:
            adminId: { type: number }
            username: { type: string }

    # ── Dashboard ─────────────────────────────────────────────────────────────────
...
    LicenseKey:
      type: object
      required: [id, keyDisplay, status, totalUses, dailyUses, maxConcurrent, concurrentSlots, createdAt, updatedAt]
      properties:
        id: { type: number }
        keyDisplay: { type: string }
        status:
          type: string
          enum: [active, revoked, locked, expired]
        expiresAt: { type: ["string", "null"] }
        maxTotalUses: { type: ["number", "null"] }
        totalUses: { type: number }
        dailyLimit: { type: ["number", "null"] }
        dailyUses: { type: number }
        maxConcurrent: { type: number }
        concurrentSlots: { type: number }
        allowedTelegramId: { type: ["string", "null"] }
        lockToTelegram: { type: boolean }
        note: { type: ["string", "null"] }
        plan:
          type: ["string", "null"]
          enum: ["basic", "pro", null]
        createdAt: { type: string }
        updatedAt: { type: string }
    KeyPage:
      type: object
      required: [keys, total, page, limit]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/LicenseKey"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
    KeyDetail:
      type: object
      required: [key]
      properties:
        key:
          $ref: "#/components/schemas/LicenseKey"
    CreateKeysInput:
      type: object
      properties:
        count: { type: number }
        durationMinutes: { type: ["number", "null"] }
        neverExpires: { type: boolean }
        maxTotalUses: { type: ["number", "null"] }
        dailyLimit: { type: ["number", "null"] }
        maxConcurrent: { type: number }
        allowedTelegramId: { type: ["string", "null"] }
        lockToTelegram: { type: boolean }
        note: { type: ["string", "null"] }
        plan:
          type: ["string", "null"]
          enum: ["basic", "pro", null]
    NewKeyRecord:
      type: object
      required: [rawKey, keyDisplay, id]
      properties:
        rawKey: { type: string }
        keyDisplay: { type: string }
        id: { type: number }
    BatchKeys:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/NewKeyRecord"
    KeyUpdateInput:
      type: object
      properties:
        action:
          type: string
          enum: [revoke, lock, unlock, extend, set_expiry]
        extraMinutes: { type: number }
        expiresAt: { type: string }
        note: { type: string }
...
    TelegramUser:
      type: object
      required: [id, telegramId, trialCount, createdAt]
      properties:
        id: { type: number }
        telegramId: { type: string }
        username: { type: ["string", "null"] }
        firstName: { type: ["string", "null"] }
        trialCount: { type: number }
        currentKeyId: { type: ["number", "null"] }
        lastUsedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    UserPage:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/TelegramUser"
        total: { type: number }
        page: { type: number }
        limit: { type: number }

    # ── Logs ──────────────────────────────────────────────────────────────────────
    UsageLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id: { type: number }
        keyId: { type: ["number", "null"] }
        telegramId: { type: ["string", "null"] }
        action: { type: string }
        ipAddress: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        createdAt: { type: string }
    UsageLogPage:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/UsageLog"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
    AuditLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id: { type: number }
        adminId: { type: ["number", "null"] }
        action: { type: string }
        targetType: { type: ["string", "null"] }
        targetId: { type: ["string", "null"] }
        ipAddress: { type: ["string", "null"] }
        createdAt: { type: string }
    AuditLogPage:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total: { type: number }
        page: { type: number }
        limit: { type: number }

    # ── Settings ──────────────────────────────────────────────────────────────────
    SystemSettings:
      type: object
      properties:
        id: { type: number }
        telegramBotToken: { type: ["string", "null"] }
        timezone: { type: ["string", "null"] }
        defaultDurationMinutes: { type: ["number", "null"] }
        defaultMaxUses: { type: ["number", "null"] }
        defaultDailyLimit: { type: ["number", "null"] }
        defaultMaxConcurrent: { type: ["number", "null"] }
        notifyExpiryDays: { type: ["number", "null"] }
        welcomeMessage: { type: ["string", "null"] }
        basicPrice: { type: ["number", "null"] }
        proPrice: { type: ["number", "null"] }
        basicStockTarget: { type: ["number", "null"] }
        proStockTarget: { type: ["number", "null"] }
        bankName: { type: ["string", "null"] }
        bankBin: { type: ["string", "null"] }
        bankAccount: { type: ["string", "null"] }
        bankHolder: { type: ["string", "null"] }
        paymentEnabled: { type: ["number", "null"] }
        updatedAt: { type: ["string", "null"] }
    SettingsEnvelope:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/SystemSettings"
    SettingsInput:
      type: object
      properties:
        telegramBotToken: { type: string }
        timezone: { type: string }
        defaultDurationMinutes: { type: number }
        defaultMaxUses: { type: ["number", "null"] }
        defaultDailyLimit: { type: ["number", "null"] }
        defaultMaxConcurrent: { type: number }
        notifyExpiryDays: { type: number }
        welcomeMessage: { type: string }
        basicPrice: { type: number }
        proPrice: { type: number }
        basicStockTarget: { type: number }
        proStockTarget: { type: number }
        bankName: { type: string }
        bankBin: { type: string }
        bankAccount: { type: string }
        bankHolder: { type: string }
        paymentEnabled: { type: boolean }
...
    Order:
      type: object
      properties:
        id: { type: number }
        telegramId: { type: string }
        username: { type: ["string", "null"] }
        plan: { type: string }
        amount: { type: number }
        orderCode: { type: string }
        status: { type: string }
        keyId: { type: ["number", "null"] }
        createdAt: { type: string }
        paidAt: { type: ["string", "null"] }
        deliveredAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
    OrdersEnvelope:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
...
    AutoStockInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [basic, pro]
    AutoStockResult:
      type: object
      required: [created, available, target]
      properties:
        created: { type: number }
        available: { type: number }
        target: { type: number }
        newAvailable: { type: number }
        message: { type: string }
        keys:
          type: array
          items:
            type: object
            properties:
              id: { type: number }
              key: { type: string }
              display: { type: string }