# lib/api-zod/src/generated/types/ordersEnvelope.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ChatGPT Account Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface OrdersEnvelope {
  orders?: Order[];
  total?: number;
  page?: number;
  limit?: number;
}